refactor(phase-4): split GitHub.swift into 3 focused files (#607)
- GitHubURLSessionTransport.swift: urlSessionAPI, urlSessionAPIPaginated,
  extractNextURL, ghAPI, ghAPIPaginated, rate-limit flag
- GitHubCLITransport.swift: runGHProcess, ghAPICLI, ghAPIPaginatedCLI,
  deleteRunnerByID, patchRunnerLabels, ghPost, cancelRun,
  fetchRegistrationToken, fetchRemovalToken, ghBinaryPath
- GitHub.swift: domain fetchers only (fetchActiveJobs, fetchRunners,
  fetchUserOrgs, fetchUserRepos, fetchStepLog, scopeFromHtmlUrl,
  runIDFromHtmlUrl) — no swiftlint:disable block
No logic changes.
