+++ Sources/RunnerBar/GitHub.swift
@@ -1,250 +1,4 @@
-// swiftlint:disable colon identifier_name opening_brace cyclomatic_complexity function_body_length multiple_closures_with_trailing_closure type_body_length
 import Foundation
-import os
-
-// MARK: - Rate limit flag
-
-private let _rateLimitLock = OSAllocatedUnfairLock(initialState: false)
-
-var ghIsRateLimited: Bool {
-    get { _rateLimitLock.withLock { $0 } }
-    set { _rateLimitLock.withLock { $0 = newValue } }
-}
-
-// MARK: - URLSession API (native, preferred when OAuth token is available)
-//
-// urlSessionAPI / urlSessionAPIPaginated use the Keychain token (set by OAuthService)
-// with a plain URLSession + Authorization: Bearer header.
-//
-// Both functions block the calling thread via DispatchSemaphore. They must
-// always be called from a background thread — a precondition assertion guards
-// this at runtime. All existing call sites dispatch via DispatchQueue.global().
-
-private let apiBase = "https://api.github.com"
-
-/// Fetches a single GitHub API page synchronously (blocking the calling thread).
-/// ⚠️ Must be called from a background thread, never from the main thread.
-func urlSessionAPI(_ endpoint: String, timeout: TimeInterval = 20) -> Data? {
-    dispatchPrecondition(condition: .notOnQueue(.main))
-    guard let token = githubToken() else {
-        log("urlSessionAPI › no token available")
-        return nil
-    }
-    let urlString = endpoint.hasPrefix("http") ? endpoint : "\(apiBase)/\(endpoint.trimmingCharacters(in: CharacterSet(charactersIn: "/")))"
-    guard let url = URL(string: urlString) else {
-        log("urlSessionAPI › invalid URL: \(urlString)")
-        return nil
-    }
-    var req = URLRequest(url: url, timeoutInterval: timeout)
-    req.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-    req.setValue("application/vnd.github+json", forHTTPHeaderField: "Accept")
-    req.setValue("2022-11-28", forHTTPHeaderField: "X-GitHub-Api-Version")
-
-    let sem = DispatchSemaphore(value: 0)
-    var result: Data?
-    URLSession.shared.dataTask(with: req) { data, response, error in
-        defer { sem.signal() }
-        if let error {
-            log("urlSessionAPI › network error: \(error.localizedDescription)")
-            return
-        }
-        if let http = response as? HTTPURLResponse {
-            log("urlSessionAPI › \(urlString) status=\(http.statusCode)")
-            if http.statusCode == 403 || http.statusCode == 429 {
-                ghIsRateLimited = true
-                return
-            }
-            guard (200..<300).contains(http.statusCode) else { return }
-        }
-        result = data
-    }.resume()
-    sem.wait()
-    return result
-}
-
-/// Fetches all pages of a GitHub API endpoint, concatenating JSON arrays.
-/// Follows the `Link: <url>; rel="next"` header automatically.
-/// ⚠️ Must be called from a background thread, never from the main thread.
-func urlSessionAPIPaginated(_ endpoint: String, timeout: TimeInterval = 60) -> Data? {
-    dispatchPrecondition(condition: .notOnQueue(.main))
-    guard let token = githubToken() else {
-        log("urlSessionAPIPaginated › no token available")
-        return nil
-    }
-    var nextURL: String? = endpoint.hasPrefix("http") ? endpoint : "\(apiBase)/\(endpoint.trimmingCharacters(in: CharacterSet(charactersIn: "/")))"
-    var allItems: [[String: Any]] = []
-
-    while let urlString = nextURL {
-        guard let url = URL(string: urlString) else { break }
-        var req = URLRequest(url: url, timeoutInterval: timeout)
-        req.setValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
-        req.setValue("application/vnd.github+json", forHTTPHeaderField: "Accept")
-        req.setValue("2022-11-28", forHTTPHeaderField: "X-GitHub-Api-Version")
-
-        let sem = DispatchSemaphore(value: 0)
-        var pageData: Data?
-        var linkHeader: String?
-        URLSession.shared.dataTask(with: req) { data, response, error in
-            defer { sem.signal() }
-            if let error {
-                log("urlSessionAPIPaginated › network error: \(error.localizedDescription)")
-                return
-            }
-            if let http = response as? HTTPURLResponse {
-                if http.statusCode == 403 || http.statusCode == 429 {
-                    ghIsRateLimited = true
-                    return
-                }
-                guard (200..<300).contains(http.statusCode) else { return }
-                linkHeader = http.value(forHTTPHeaderField: "Link")
-            }
-            pageData = data
-        }.resume()
-        sem.wait()
-
-        guard let data = pageData else { break }
-        if let page = try? JSONSerialization.jsonObject(with: data) as? [[String: Any]] {
-            allItems.append(contentsOf: page)
-        }
-        nextURL = extractNextURL(from: linkHeader)
-    }
-
-    guard !allItems.isEmpty else { return nil }
-    return try? JSONSerialization.data(withJSONObject: allItems)
-}
-
-/// Parses the `Link` header and returns the URL for `rel="next"`, if present.
-/// Accepts entries with extra attributes (e.g. rel="next"; title="next page").
-private func extractNextURL(from header: String?) -> String? {
-    guard let header else { return nil }
-    for part in header.components(separatedBy: ",") {
-        let segments = part.components(separatedBy: ";")
-        // Need at least a link segment and one attribute segment.
-        guard segments.count >= 2 else { continue }
-        let linkSegment = segments[0].trimmingCharacters(in: .whitespaces)
-        guard linkSegment.hasPrefix("<"), linkSegment.hasSuffix(">") else { continue }
-        let link = String(linkSegment.dropFirst().dropLast())
-        // Check all attribute segments for rel="next".
-        let isNext = segments.dropFirst().contains(where: {
-            $0.trimmingCharacters(in: .whitespaces) == "rel=\"next\""
-        })
-        if isNext { return link }
-    }
-    return nil
-}
-
-// MARK: - Public API entry points
-//
-// Prefer URLSession (native OAuth token) when available; fall back to gh CLI subprocess.
-
-func ghAPI(_ endpoint: String, timeout: TimeInterval = 20) -> Data? {
-    if githubToken() != nil {
-        let data = urlSessionAPI(endpoint, timeout: timeout)
-        if data != nil { return data }
-    }
-    return ghAPICLI(endpoint, timeout: timeout)
-}
-
-func ghAPIPaginated(_ endpoint: String, timeout: TimeInterval = 60) -> Data? {
-    if githubToken() != nil {
-        let data = urlSessionAPIPaginated(endpoint, timeout: timeout)
-        if data != nil { return data }
-    }
-    return ghAPIPaginatedCLI(endpoint, timeout: timeout)
-}
-
-// MARK: - gh CLI subprocess (fallback)
-
-private func runGHProcess(arguments: [String], timeout: TimeInterval = 20) -> Data? {
-    guard let ghPath = ghBinaryPath() else {
-        log("runGHProcess › gh not found in known paths")
-        return nil
-    }
-    log("runGHProcess › \(ghPath) \(arguments.joined(separator: " "))")
-    let task = Process()
-    let pipe = Pipe()
-    task.executableURL = URL(fileURLWithPath: ghPath)
-    task.arguments = arguments
-    task.standardOutput = pipe
-    task.standardError = Pipe()
-    var outputData = Data()
-    let lock = NSLock()
-    pipe.fileHandleForReading.readabilityHandler = { handle in
-        let chunk = handle.availableData
-        guard !chunk.isEmpty else { return }
-        lock.lock(); outputData.append(chunk); lock.unlock()
-    }
-    do { try task.run() } catch {
-        log("runGHProcess › launch error: \(error)")
-        pipe.fileHandleForReading.readabilityHandler = nil
-        return nil
-    }
-    let timeoutItem = DispatchWorkItem { task.terminate() }
-    DispatchQueue.global().asyncAfter(deadline: .now() + timeout, execute: timeoutItem)
-    task.waitUntilExit()
-    timeoutItem.cancel()
-    pipe.fileHandleForReading.readabilityHandler = nil
-    let tail = pipe.fileHandleForReading.readDataToEndOfFile()
-    if !tail.isEmpty { lock.lock(); outputData.append(tail); lock.unlock() }
-    log("runGHProcess › exit=\(task.terminationStatus) bytes=\(outputData.count)")
-    return outputData.isEmpty ? nil : outputData
-}
-
-private func ghAPICLI(_ endpoint: String, timeout: TimeInterval = 20) -> Data? {
-    guard let outputData = runGHProcess(arguments: ["api", endpoint], timeout: timeout) else {
-        return nil
-    }
-    log("ghAPICLI › \(endpoint) → \(outputData.count)b")
-    if let json = try? JSONSerialization.jsonObject(with: outputData) as? [String: Any],
-       let status = json["status"] as? String,
-       status == "403" || status == "429" {
-        ghIsRateLimited = true
-        log("ghAPICLI › rate limit (\(status)): \(endpoint)")
-        return nil
-    }
-    return outputData
-}
-
-private func ghAPIPaginatedCLI(_ endpoint: String, timeout: TimeInterval = 60) -> Data? {
-    guard let ghPath = ghBinaryPath() else { log("ghAPIPaginatedCLI › gh not found"); return nil }
-    let task = Process()
-    let pipe = Pipe()
-    task.executableURL = URL(fileURLWithPath: ghPath)
-    task.arguments = ["api", "--paginate", endpoint]
-    task.standardOutput = pipe
-    task.standardError = Pipe()
-    var outputData = Data()
-    let lock = NSLock()
-    pipe.fileHandleForReading.readabilityHandler = { handle in
-        let chunk = handle.availableData
-        guard !chunk.isEmpty else { return }
-        lock.lock(); outputData.append(chunk); lock.unlock()
-    }
-    do { try task.run() } catch {
-        log("ghAPIPaginatedCLI › launch error: \(error)")
-        pipe.fileHandleForReading.readabilityHandler = nil
-        return nil
-    }
-    let timeoutItem = DispatchWorkItem { task.terminate() }
-    DispatchQueue.global().asyncAfter(deadline: .now() + timeout, execute: timeoutItem)
-    task.waitUntilExit()
-    timeoutItem.cancel()
-    pipe.fileHandleForReading.readabilityHandler = nil
-    let tail = pipe.fileHandleForReading.readDataToEndOfFile()
-    if !tail.isEmpty { lock.lock(); outputData.append(tail); lock.unlock() }
-    log("ghAPIPaginatedCLI › \(endpoint) → \(outputData.count)b exit \(task.terminationStatus)")
-    if task.terminationStatus != 0 {
-        let raw = String(data: outputData, encoding: .utf8) ?? ""
-        if raw.contains("\"403\"") || raw.contains("\"429\"") || raw.contains("rate limit") {
-            ghIsRateLimited = true
-            log("ghAPIPaginatedCLI › rate limit detected: \(endpoint)")
-        } else {
-            log("ghAPIPaginatedCLI › non-zero exit (\(task.terminationStatus)): \(endpoint)")
-        }
-        return nil
-    }
-    return outputData.isEmpty ? nil : outputData
-}
 
 // MARK: - URL helpers
 
@@ -313,12 +67,9 @@ private struct WorkflowRun: Codable { let id: Int }
 // MARK: - Runners
 
 func fetchRunners(for scope: String) -> [Runner] {
-    let endpoint: String
-    if scope.contains("/") {
-        endpoint = "repos/\(scope)/actions/runners"
-    } else {
-        endpoint = "orgs/\(scope)/actions/runners"
-    }
+    let endpoint = scope.contains("/")
+        ? "repos/\(scope)/actions/runners"
+        : "orgs/\(scope)/actions/runners"
     log("fetchRunners › \(endpoint)")
     guard let data = ghAPI(endpoint) else {
         log("fetchRunners › no data for scope: \(scope)")
@@ -350,178 +101,6 @@ func fetchUserRepos() -> [String] {
     }
     guard let repos = try? JSONDecoder().decode([Repo].self, from: data) else { return [] }
     return repos.map(\.fullName)
-}
-
-// MARK: - Registration token
-
-/// Fetches a short-lived runner registration token for the given scope.
-/// Required by `config.sh --token` during runner setup.
-/// NOTE: raw response is intentionally not logged — it contains a short-lived secret.
-func fetchRegistrationToken(scope: String) -> String? {
-    let endpoint: String
-    if scope.contains("/") {
-        endpoint = "repos/\(scope)/actions/runners/registration-token"
-    } else {
-        endpoint = "orgs/\(scope)/actions/runners/registration-token"
-    }
-    log("fetchRegistrationToken › POSTing \(endpoint)")
-    let args = ["api", "--method", "POST",
-                "-H", "Accept: application/vnd.github+json", endpoint]
-    guard let outputData = runGHProcess(arguments: args, timeout: 30) else {
-        log("fetchRegistrationToken › no data for \(endpoint)")
-        return nil
-    }
-    log("fetchRegistrationToken › \(endpoint) \(outputData.count)b raw=[REDACTED]")
-    struct TokenResponse: Decodable { let token: String }
-    guard let resp = try? JSONDecoder().decode(TokenResponse.self, from: outputData) else {
-        log("fetchRegistrationToken › decode failed for \(endpoint) (\(outputData.count)b)")
-        return nil
-    }
-    log("fetchRegistrationToken › got token (first 4): \(resp.token.prefix(4))...")
-    return resp.token
-}
-
-// MARK: - Removal token
-
-/// Fetches a runner removal token for the given scope (owner/repo or org).
-/// Required by `config.sh remove --token <token>`.
-/// NOTE: raw response is intentionally not logged — it contains a short-lived secret.
-func fetchRemovalToken(scope: String) -> String? {
-    let endpoint: String
-    if scope.contains("/") {
-        endpoint = "repos/\(scope)/actions/runners/remove-token"
-    } else {
-        endpoint = "orgs/\(scope)/actions/runners/remove-token"
-    }
-    log("fetchRemovalToken › POSTing \(endpoint)")
-    let args = ["api", "--method", "POST",
-                "-H", "Accept: application/vnd.github+json", endpoint]
-    guard let outputData = runGHProcess(arguments: args, timeout: 30) else {
-        log("fetchRemovalToken › no data returned for \(endpoint)")
-        return nil
-    }
-    log("fetchRemovalToken › raw response (\(outputData.count)b): [REDACTED]")
-    struct TokenResponse: Decodable { let token: String }
-    guard let resp = try? JSONDecoder().decode(TokenResponse.self, from: outputData) else {
-        log("fetchRemovalToken › decode failed for \(endpoint) (\(outputData.count)b)")
-        return nil
-    }
-    log("fetchRemovalToken › got removal token (first 4): \(resp.token.prefix(4))...")
-    return resp.token
-}
-
-// MARK: - Delete runner by ID (API fallback for corrupt installs)
-
-/// Directly deregisters a runner from GitHub via DELETE API.
-/// Used as fallback when config.sh is missing or corrupt.
-/// Returns true only if the runner was successfully deleted (HTTP 204, gh exit 0).
-/// A 404 response means the runner ID was not found on GitHub — that is a failure,
-/// not a deletion. Non-zero exit = runner may still exist on GitHub.
-@discardableResult
-func deleteRunnerByID(scope: String, runnerID: Int) -> Bool {
-    let endpoint: String
-    if scope.contains("/") {
-        endpoint = "repos/\(scope)/actions/runners/\(runnerID)"
-    } else {
-        endpoint = "orgs/\(scope)/actions/runners/\(runnerID)"
-    }
-    log("deleteRunnerByID › DELETE \(endpoint) runnerID=\(runnerID)")
-    guard let ghPath = ghBinaryPath() else {
-        log("deleteRunnerByID › gh not found")
-        return false
-    }
-    let task = Process()
-    let pipe = Pipe()
-    task.executableURL = URL(fileURLWithPath: ghPath)
-    task.arguments = ["api", "--method", "DELETE",
-                      "-H", "Accept: application/vnd.github+json", endpoint]
-    task.standardOutput = pipe
-    task.standardError = pipe
-    do { try task.run() } catch {
-        log("deleteRunnerByID › launch error: \(error)")
-        return false
-    }
-    let timeoutItem = DispatchWorkItem { task.terminate() }
-    DispatchQueue.global().asyncAfter(deadline: .now() + 30, execute: timeoutItem)
-    task.waitUntilExit()
-    timeoutItem.cancel()
-    let outData = pipe.fileHandleForReading.readDataToEndOfFile()
-    let raw = String(data: outData, encoding: .utf8) ?? ""
-    let status = task.terminationStatus
-    log("deleteRunnerByID › exit=\(status) response=\(raw.prefix(200))")
-    let ok = status == 0
-    if !ok {
-        log("deleteRunnerByID › DELETE failed (exit=\(status)) for runnerID=\(runnerID) — runner may still exist on GitHub")
-    }
-    log("deleteRunnerByID › result=\(ok) for runnerID=\(runnerID)")
-    return ok
-}
-
-// MARK: - Patch runner labels (#492)
-
-/// Replaces ALL custom labels on the runner identified by `runnerID` within `scope`.
-/// Built-in labels (self-hosted, OS, arch) are preserved by GitHub automatically.
-/// Returns the updated label names on success, or nil on failure.
-@discardableResult
-func patchRunnerLabels(scope: String, runnerID: Int, labels: [String]) -> [String]? {
-    let endpoint: String
-    if scope.contains("/") {
-        endpoint = "repos/\(scope)/actions/runners/\(runnerID)/labels"
-    } else {
-        endpoint = "orgs/\(scope)/actions/runners/\(runnerID)/labels"
-    }
-    log("patchRunnerLabels › PUT \(endpoint) labels=\(labels)")
-    guard let bodyData = try? JSONSerialization.data(withJSONObject: ["labels": labels]),
-          let bodyString = String(data: bodyData, encoding: .utf8),
-          let ghPath = ghBinaryPath()
-    else {
-        log("patchRunnerLabels › failed to build request")
-        return nil
-    }
-    let task = Process()
-    let pipe = Pipe()
-    let errPipe = Pipe()
-    task.executableURL = URL(fileURLWithPath: ghPath)
-    task.arguments = [
-        "api",
-        "--method", "PUT",
-        "-H", "Accept: application/vnd.github+json",
-        "-H", "Content-Type: application/json",
-        "--input", "-",
-        endpoint
-    ]
-    task.standardOutput = pipe
-    task.standardError = errPipe
-    let inputPipe = Pipe()
-    task.standardInput = inputPipe
-    do { try task.run() } catch {
-        log("patchRunnerLabels › launch error: \(error)")
-        return nil
-    }
-    inputPipe.fileHandleForWriting.write(bodyData)
-    inputPipe.fileHandleForWriting.closeFile()
-    let timeoutItem = DispatchWorkItem { task.terminate() }
-    DispatchQueue.global().asyncAfter(deadline: .now() + 30, execute: timeoutItem)
-    task.waitUntilExit()
-    timeoutItem.cancel()
-    let outData = pipe.fileHandleForReading.readDataToEndOfFile()
-    let raw = String(data: outData, encoding: .utf8) ?? ""
-    log("patchRunnerLabels › exit=\(task.terminationStatus) response=\(raw.prefix(300))")
-    guard task.terminationStatus == 0 else {
-        log("patchRunnerLabels › non-zero exit for endpoint=\(endpoint) body=\(bodyString)")
-        return nil
-    }
-    struct LabelsResponse: Decodable {
-        struct Label: Decodable { let name: String }
-        let labels: [Label]
-    }
-    guard let resp = try? JSONDecoder().decode(LabelsResponse.self, from: outData) else {
-        log("patchRunnerLabels › decode failed raw=\(raw.prefix(200))")
-        return nil
-    }
-    let names = resp.labels.map(\.name)
-    log("patchRunnerLabels › success labels=\(names)")
-    return names
 }
 
 // MARK: - Step log
@@ -586,45 +165,3 @@ private func stripAnsi(_ input: String) -> String {
         withTemplate: ""
     )
 }
-
-// MARK: - Shared gh binary path
-
-func ghBinaryPath() -> String? {
-    let candidates = ["/opt/homebrew/bin/gh", "/usr/local/bin/gh", "/usr/bin/gh"]
-    let found = candidates.first(where: { FileManager.default.isExecutableFile(atPath: $0) })
-    if found == nil { log("ghBinaryPath › gh not found in \(candidates)") }
-    return found
-}
-
-// MARK: - POST helper
-
-@discardableResult
-func ghPost(_ endpoint: String) -> Bool {
-    guard let ghPath = ghBinaryPath() else { log("ghPost › gh not found"); return false }
-    let task = Process()
-    task.executableURL = URL(fileURLWithPath: ghPath)
-    task.arguments = ["api", "--method", "POST",
-                      "-H", "Accept: application/vnd.github+json", endpoint]
-    task.standardOutput = Pipe()
-    task.standardError = Pipe()
-    do { try task.run() } catch {
-        log("ghPost › launch error: \(error)")
-        return false
-    }
-    let timeoutItem = DispatchWorkItem { task.terminate() }
-    DispatchQueue.global().asyncAfter(deadline: .now() + 30, execute: timeoutItem)
-    task.waitUntilExit()
-    timeoutItem.cancel()
-    log("ghPost › \(endpoint) exit \(task.terminationStatus)")
-    return task.terminationStatus == 0
-}
-
-// MARK: - Cancel run
-
-@discardableResult
-func cancelRun(runID: Int, scope: String) -> Bool {
-    let result = ghPost("repos/\(scope)/actions/runs/\(runID)/cancel")
-    log("cancelRun › run=\(runID) scope=\(scope) success=\(result)")
-    return result
-}
-// swiftlint:enable colon identifier_name opening_brace cyclomatic_complexity function_body_length multiple_closures_with_trailing_closure type_body_length
